+++ include/color.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   color.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/25 23:17:02 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 16:13:45 by mkhan            ###   ########.fr       */
+/*   Updated: 2023/01/20 18:55:51 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,19 +23,19 @@
 typedef struct s_color		t_color;
 struct s_color
 {
-	float	r;
-	float	g;
-	float	b;
-	float	a;
+	double	r;
+	double	g;
+	double	b;
+	double	a;
 };
 
 unsigned int	create_mlx_color(t_color *color);
 void			add_colors(t_color *res, const t_color *c1, const t_color *c2);
-void			mult_color(t_color *res, const t_color *color, float val);
+void			mult_color(t_color *res, const t_color *color, double val);
 void			blend_colors(t_color *res, const t_color *c1,
 					const t_color *c2);
 
-int				color_mix(int c1, int c2, float mix);
+int				color_mix(int c1, int c2, double mix);
 int				color_difference(int c1, int c2);
 void			sub_colors(t_color *res, const t_color *c1, const t_color *c2);
