+++ libft/ft_swap.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/16 16:13:40 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/20 18:55:51 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ void	ft_swap(int *a, int *b)
 	*b = tmp;
 }
 
-void	ft_swapd(float *a, float *b)
+void	ft_swapd(double *a, double *b)
 {
-	float	tmp;
+	double	tmp;
 
 	tmp = *a;
 	*a = *b;
