+++ src/math/vector_operations.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 09:54:26 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/20 18:55:51 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
  * @param vec A vector
  * @return vec's magnitude
  */
-float	vec_magnitude(const t_vector *vec)
+double	vec_magnitude(const t_vector *vec)
 {
 	return (sqrt(vec->x * vec->x + vec->y * vec->y \
 			+ vec->z * vec->z));
@@ -39,7 +39,7 @@ void	normalize_vec(t_vector *vec)
  * @param v2 Second vector
  * @return The dot product of v1 and v2
  */
-float	dot_product(const t_vector *v1, const t_vector *v2)
+double	dot_product(const t_vector *v1, const t_vector *v2)
 {
 	return (v1->x * v2->x + v1->y * v2->y + v1->z * v2->z + v1->w * v2->w);
 }
@@ -59,7 +59,7 @@ void	cross_product(t_vector *res, const t_vector *v1, const t_vector *v2)
 	res->w = 0;
 }
 
-float	vec_distance(const t_vector *v1, const t_vector *v2)
+double	vec_distance(const t_vector *v1, const t_vector *v2)
 {
 	t_vector	sub;
