+++ src/parsing/settings_validation.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/06 16:19:56 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/18 18:00:13 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/20 18:55:51 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,10 +66,10 @@ bool	is_valid_color(const char *color)
 	return (false);
 }
 
-bool	check_value(const char *key, const char *val, float min, float max)
+bool	check_value(const char *key, const char *val, double min, double max)
 {
 	bool	success;
-	float	parsed_value;
+	double	parsed_value;
 
 	success = true;
 	if (ft_strcmp_case(key, "diffuse_texture") == 0 || ft_strcmp_case(key, "normal_texture") == 0)
@@ -99,8 +99,8 @@ bool	check_value(const char *key, const char *val, float min, float max)
 
 bool	is_valid_val(const char *key, const char *val)
 {
-	float	min;
-	float	max;
+	double	min;
+	double	max;
 
 	if (val == NULL || ft_strlen(val) == 0)
 		return (false);
