+++ src/renderer/camera.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/02 17:39:44 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/10 15:42:31 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/20 18:55:51 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,10 +63,10 @@ void	view_transform(t_mat4 *res, const t_vector *from, const t_vector *up,
 
 void	camera_init(t_camera *camera, t_scene *scene)
 {
-	float	half_view;
-	float	aspect;
-	float	h;
-	float	w;
+	double	half_view;
+	double	aspect;
+	double	h;
+	double	w;
 
 	h = scene->settings.render_h;
 	w = scene->settings.render_w;
